incap.X measuring freq to 1000th place, average needed at high frequncies(uneccessary for us?)

+++ incap.X/include/INPUT_CAPTURE.h
@@ -5,4 +5,4 @@
 
 char INCAP_Init(void);
 
-float INCAP_getFreq(void);
+unsigned int INCAP_getFreq(void);
